dbeaver/dbeaver-ee#1529 fix eula dialog not being shown after license (#16160)
* dbeaver/dbeaver-ee#1529 fix eula dialog not being shown after license & refactoring

* dbeaver/dbeaver-ee#1529 init workbench Initializers after license dialog & refactoring

* dbeaver/dbeaver-ee#1529 add priority parameter to xml & use TreeMap to sort by priority
