+++ plugins/org.jkiss.dbeaver.core/META-INF/MANIFEST.MF
@@ -18,6 +18,7 @@ Export-Package: org.jkiss.dbeaver,
  org.jkiss.dbeaver.ui.actions.datasource,
  org.jkiss.dbeaver.ui.dialogs.connection,
  org.jkiss.dbeaver.ui.e4,
+ org.jkiss.dbeaver.ui.eula,
  org.jkiss.dbeaver.ui.perspective,
  org.jkiss.dbeaver.ui.preferences,
  org.jkiss.dbeaver.ui.resources,
